app/requirements.txt
fastapi==0.115.8
uvicorn==0.34.0
SQLAlchemy==2.0.36
python-dotenv==1.0.1
psycopg2==2.9.10
pydantic==2.10.3